pyproject.toml
[tool.poetry]
name = "ozon-clone-backend"
version = "0.1.0"
description = ""
authors = ["Rashid Mamarimov <sonochiwaxd@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "ozon-clone-backend" }]

[tool.poetry.dependencies]
python = "3.11.4"
uvicorn = "^0.22.0"
alembic = "^1.11.1"
sqlalchemy = "^2.0.15"
fastapi-users = {extras = ["sqlalchemy"], version = "^12.1.0"}
celery = "^5.2.7"
pydantic = "2.0.2"
httpx = "^0.24.1"
pytest = "^7.4.0"
flower = "^2.0.0"
fastapi = "0.100"
pydantic-settings = "^2.0.1"
redis = "^4.6.0"
itsdangerous = "^2.1.2"
psycopg = {extras = ["binary"], version = "^3.1.9"}
python-multipart = "^0.0.6"
python-slugify = "^8.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
